ESIMW-1669: Fix issue with static map reference (#69)
This code was making a static copy of a constant map which it would clear out and repopulate each time the value was referenced.  It was previously referencing a modifiable map which would cause it to clear out a "constant" value and then the map would remain empty until the VM was restarted.  This removes the second copy of the map and references the original since there does not appear to be a need for the extra copy.