ESIMW-2253: Encode URL parameter keys (#71)
There was an issue where Rice was not escaping square brackets in URLS which causes issues with a newer version of Tomcat that requires those be specified in an encoding whitelist. This patch fixes the issue within Rice eliminating the need for changes to Tomcat configurations.

See https://bz.apache.org/bugzilla/show_bug.cgi?id=60594  for more information.